+++ lib/libvfio-user.c
@@ -1905,8 +1905,8 @@ vfu_setup_region(vfu_ctx_t *vfu_ctx, int region_idx, size_t size,
         struct iovec *iov = &mmap_areas[i];
         if ((size_t)iov_end(iov) > size) {
             vfu_log(vfu_ctx, LOG_ERR, "mmap area #%zu %#llx-%#llx exceeds region size of %#llx\n",
-                    i, (unsigned long long)iov->iov_base,
-		    (unsigned long long)(iov->iov_base) + iov->iov_len - 1,
+                    i, (unsigned long long)(uintptr_t)iov->iov_base,
+		    (unsigned long long)(uintptr_t)(iov->iov_base) + iov->iov_len - 1,
 		    (unsigned long long)size);
             return ERROR_INT(EINVAL);
         }
